requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
cffi==1.14.0
click==7.1.1
Flask==1.1.1
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycparser==2.20
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
Werkzeug==1.0.0